fix(config): replace "{id}" with `make_random_file_stem()`
